# contracts/openapi/ledger-v1.yaml
openapi: 3.1.0
info:
  title: Ledger API
  version: 1.0.0
  description: >-
    Accounts, holds and the double-entry journal. The ledger service serves this contract, and its
    integration tests check every response against this file.
servers:
  - url: http://localhost:8080/ledger
    description: Through the local gateway
tags:
  - name: Accounts
  - name: Holds
  - name: Journal
security:
  - bearerAuth: []
paths:
  /v1/accounts:
    post:
      operationId: openAccount
      summary: Open an account
      description: >-
        The account is ACTIVE at once. Sending the same request again with the same Idempotency-Key
        returns the first account. The same key with a different body is refused with 422.
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: The account was opened
          headers:
            Location:
              description: URL of the new account
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /v1/accounts/{accountId}:
    get:
      operationId: getAccount
      summary: Get one account
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: The account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/accounts/{accountId}/balance:
    get:
      operationId: getBalance
      summary: Get the balances of an account
      description: >-
        The ledger balance counts posted entries. The held amount is reserved by holds. The available
        balance is the ledger balance minus the held amount.
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: The balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/accounts/{accountId}/entries:
    get:
      operationId: listEntries
      summary: List the journal entries that touched an account, newest first
      tags: [Journal]
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: One page of entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryPage'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/accounts/{accountId}/freeze:
    post:
      operationId: freezeAccount
      summary: Freeze an account
      description: A frozen account still receives money. It cannot send any, and no new hold can be placed on it.
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The account, now FROZEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/accounts/{accountId}/unfreeze:
    post:
      operationId: unfreezeAccount
      summary: Lift a freeze
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The account, ACTIVE again
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/accounts/{accountId}/close:
    post:
      operationId: closeAccount
      summary: Close an account
      description: Only an account with a zero balance and no holds can be closed. Closing cannot be undone.
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The account, now CLOSED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/holds:
    post:
      operationId: placeHold
      summary: Reserve money on an account
      description: >-
        Internal. The reserved money stays in the account but cannot be spent by anything else. A
        reference holds money once. Asking again with the same reference returns the first hold with 200
        while it is still active, and 409 once it has been released or captured.
      tags: [Holds]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceHoldRequest'
      responses:
        '201':
          description: The money was reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hold'
        '200':
          description: This reference already had a hold, which is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hold'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /v1/holds/{holdId}:
    get:
      operationId: getHold
      summary: Get one hold
      tags: [Holds]
      parameters:
        - $ref: '#/components/parameters/HoldId'
      responses:
        '200':
          description: The hold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hold'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/holds/{holdId}/release:
    post:
      operationId: releaseHold
      summary: Give reserved money back
      description: Internal. Releasing a hold that is already released changes nothing and returns it.
      tags: [Holds]
      parameters:
        - $ref: '#/components/parameters/HoldId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The hold, now RELEASED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hold'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/journal-entries:
    post:
      operationId: postJournalEntry
      summary: Post a balanced journal entry
      description: >-
        Internal. Called by other services, never by a customer. Debits must equal credits. A reference
        is posted once. Posting it again with the same lines returns the first entry with 200, and
        with different lines is refused with 409. With a holdId, the hold is ended in the same
        transaction, so the reserved money is spent by this entry and by nothing else.
      tags: [Journal]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostJournalEntryRequest'
      responses:
        '201':
          description: The entry was posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
        '200':
          description: This reference was already posted with the same lines. The first entry is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: The same key on a retry returns the first response again.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._:-]{8,64}$'
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    HoldId:
      name: holdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      description: The `nextCursor` value of the previous page.
      schema:
        type: string
  responses:
    Problem:
      description: An error in RFC 9457 problem details format
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?[0-9]+([.][0-9]+)?$'
          examples: ['125.50']
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          examples: [USD]
    AccountType:
      type: string
      enum: [CHECKING, SAVINGS, INTERNAL]
    AccountStatus:
      type: string
      enum: [PENDING, ACTIVE, FROZEN, CLOSED]
    HoldStatus:
      type: string
      description: >-
        ACTIVE reserves money. CAPTURED was spent by a journal entry. RELEASED gave it back, because
        someone asked or because the hold ran out of time.
      enum: [ACTIVE, CAPTURED, RELEASED]
    OpenAccountRequest:
      type: object
      required: [customerId, type, currency]
      properties:
        customerId:
          type: string
          format: uuid
        type:
          type: string
          enum: [CHECKING, SAVINGS]
          description: >-
            Only customer accounts are opened here. The bank's own INTERNAL accounts have no overdraft
            limit, so they are created by a database migration and never through the API.
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        name:
          type: string
          maxLength: 140
    Account:
      type: object
      required: [id, type, currency, status, openedAt]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AccountType'
        name:
          type: string
        currency:
          type: string
        status:
          $ref: '#/components/schemas/AccountStatus'
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    Balance:
      type: object
      required: [accountId, ledger, available, held, asOf]
      properties:
        accountId:
          type: string
          format: uuid
        ledger:
          $ref: '#/components/schemas/Money'
        available:
          $ref: '#/components/schemas/Money'
        held:
          $ref: '#/components/schemas/Money'
        asOf:
          type: string
          format: date-time
    PlaceHoldRequest:
      type: object
      required: [accountId, amount, reference]
      properties:
        accountId:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        reference:
          type: string
          maxLength: 140
          description: The business reason, such as a payment id. One reference holds money at most once.
        expiresInSeconds:
          type: integer
          minimum: 1
          maximum: 2592000
          description: After this long a hold nobody used is given back. Seven days when left out.
    Hold:
      type: object
      required: [id, accountId, amount, reference, status, createdAt, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        reference:
          type: string
        status:
          $ref: '#/components/schemas/HoldStatus'
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    Posting:
      type: object
      required: [accountId, side, amount]
      properties:
        accountId:
          type: string
          format: uuid
        side:
          type: string
          enum: [DEBIT, CREDIT]
        amount:
          $ref: '#/components/schemas/Money'
    PostJournalEntryRequest:
      type: object
      required: [reference, postings]
      properties:
        reference:
          type: string
          maxLength: 140
          description: The business reason, such as a payment id. One reference is posted at most once.
        postings:
          type: array
          minItems: 2
          maxItems: 200
          items:
            $ref: '#/components/schemas/Posting'
        holdId:
          type: string
          format: uuid
          description: A hold on one of the accounts in this entry, to end in the same transaction.
    JournalEntry:
      type: object
      required: [id, reference, postedAt, postings]
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        postedAt:
          type: string
          format: date-time
        postings:
          type: array
          items:
            $ref: '#/components/schemas/Posting'
    EntryPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntry'
        nextCursor:
          type: string
          description: Missing on the last page.
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: A stable address for this kind of error.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          description: The path of the request that failed.
        code:
          type: string
          description: A stable code a client can branch on.
          examples: [INSUFFICIENT_FUNDS]
        fields:
          type: object
          description: For a body that failed validation, each bad field with what is wrong with it.
          additionalProperties:
            type: string
        traceId:
          type: string
          description: Quote this to support. It finds the request in the traces and the logs.